(PC-12296)[API] refactor: move dms email in common file
